refactor: rename validationjs to validationConstants
